Move watchdog to uaprocessor
