+++ asyncua/server/uaprocessor.py
@@ -1,8 +1,10 @@
+import asyncio
 import copy
 import time
 import logging
 from typing import Deque, Optional, Dict
 from collections import deque
+from datetime import datetime, timedelta, timezone
 
 from asyncua import ua
 from ..ua.ua_binary import nodeid_from_binary, struct_from_binary, struct_to_binary, uatcp_to_binary
@@ -31,6 +33,7 @@ class UaProcessor:
         self.name = transport.get_extra_info('peername')
         self.sockname = transport.get_extra_info('sockname')
         self.session: Optional[InternalSession] = None
+        self.session_last_activity: Optional[datetime] = None
         self._transport = transport
         # deque for Publish Requests
         self._publish_requests: Deque[PublishRequestData] = deque()
@@ -39,6 +42,9 @@ class UaProcessor:
         self._publish_results_subs: Dict[ua.IntegerId, bool] = {}
         self._limits = copy.deepcopy(limits)  # Copy limits because they get overriden
         self._connection = SecureConnection(ua.SecurityPolicy(), self._limits)
+        self._closing: bool = False
+        self._session_watchdog_task: Optional[asyncio.Task] = None
+        self._watchdog_interval: float = 1.0
 
     def set_policies(self, policies):
         self._connection.set_policy_factories(policies)
@@ -178,6 +184,8 @@ class UaProcessor:
                 if self._connection.security_policy.permissions.check_validity(user, typeid, body) is False:
                     raise ua.uaerrors.BadUserAccessDenied
 
+        self.session_last_activity = datetime.now(timezone.utc)
+
         if typeid == ua.NodeId(ua.ObjectIds.CreateSessionRequest_Encoding_DefaultBinary):
             _logger.info("Create session request (%s)", user)
             params = struct_from_binary(ua.CreateSessionParameters, body)
@@ -185,6 +193,8 @@ class UaProcessor:
             self.session = self.iserver.create_session(self.name, external=True)
             # get a session creation result to send back
             sessiondata = await self.session.create_session(params, sockname=self.sockname)
+            self._closing = False
+            self._session_watchdog_task = asyncio.create_task(self._session_watchdog_loop())
             response = ua.CreateSessionResponse()
             response.Parameters = sessiondata
             response.Parameters.ServerCertificate = self._connection.security_policy.host_certificate
@@ -202,6 +212,7 @@ class UaProcessor:
             _logger.info("Close session request (%s)", user)
             if self.session:
                 deletesubs = ua.ua_binary.Primitives.Boolean.unpack(body)
+                self._closing = True
                 await self.session.close_session(deletesubs)
             else:
                 _logger.info("Request to close non-existing session (%s)", user)
@@ -507,5 +518,18 @@ class UaProcessor:
         everything we should
         """
         _logger.info("Cleanup client connection: %s", self.name)
+        self._closing = True
         if self.session:
             await self.session.close_session(True)
+
+    async def _session_watchdog_loop(self):
+        """
+        Checks if the session is alive
+        """
+        timeout = min(self.session.session_timeout / 2, self._watchdog_interval)
+        timeout_timedelta = timedelta(seconds=self.session.session_timeout)
+        while not self._closing:
+            await asyncio.sleep(timeout)
+            if (datetime.now(timezone.utc) - timeout_timedelta) > self.session_last_activity:
+                _logger.warning('Session timed out after %ss of inactivity', timeout_timedelta.total_seconds())
+                await self.close()
